requirements.txt
woocommerce
python-resize-image
Pillow
boto3
python-decouple